Extract manage overview model
